Refactor to 'providers' module
