+++ src/lib.rs
@@ -6,6 +6,7 @@ mod constants;
 mod http;
 mod memory;
 mod metrics;
+mod providers;
 mod signature;
 mod types;
 mod util;
@@ -17,6 +18,7 @@ pub use crate::constants::*;
 pub use crate::http::*;
 pub use crate::memory::*;
 pub use crate::metrics::*;
+pub use crate::providers::*;
 pub use crate::signature::*;
 pub use crate::types::*;
 pub use crate::util::*;
