+++ src/main.rs
@@ -61,78 +61,19 @@ fn get_providers() -> Vec<ProviderView> {
 #[update(guard = "require_register_provider")]
 #[candid_method]
 fn register_provider(provider: RegisterProvider) -> u64 {
-    let parsed_url = url::Url::parse(&provider.base_url).expect("unable to parse service_url");
-    let host = parsed_url.host_str().expect("service_url host missing");
-    validate_base_url(host);
-    validate_credential_path(&provider.credential_path);
-    let provider_id = METADATA.with(|m| {
-        let mut metadata = m.borrow().get().clone();
-        metadata.next_provider_id += 1;
-        m.borrow_mut().set(metadata.clone()).unwrap();
-        metadata.next_provider_id - 1
-    });
-    PROVIDERS.with(|p| {
-        p.borrow_mut().insert(
-            provider_id,
-            Provider {
-                provider_id,
-                owner: ic_cdk::caller(),
-                chain_id: provider.chain_id,
-                base_url: provider.base_url,
-                credential_path: provider.credential_path,
-                cycles_per_call: provider.cycles_per_call,
-                cycles_per_message_byte: provider.cycles_per_message_byte,
-                cycles_owed: 0,
-                active: true,
-            },
-        )
-    });
-    provider_id
-}
-
-#[update(guard = "require_register_provider")]
-#[candid_method]
-fn update_provider(update: UpdateProvider) {
-    PROVIDERS.with(|p| match p.borrow_mut().get(&update.provider_id) {
-        Some(mut provider) => {
-            if provider.owner != ic_cdk::caller() && !is_authorized(Auth::Admin) {
-                ic_cdk::trap("Provider owner != caller");
-            }
-            if let Some(url) = update.base_url {
-                validate_base_url(&url);
-                provider.base_url = url;
-            }
-            if let Some(path) = update.credential_path {
-                validate_credential_path(&path);
-                provider.credential_path = path;
-            }
-            if let Some(active) = update.active {
-                provider.active = active;
-            }
-            if let Some(cycles_per_call) = update.cycles_per_call {
-                provider.cycles_per_call = cycles_per_call;
-            }
-            if let Some(cycles_per_message_byte) = update.cycles_per_message_byte {
-                provider.cycles_per_message_byte = cycles_per_message_byte;
-            }
-            p.borrow_mut().insert(update.provider_id, provider);
-        }
-        None => ic_cdk::trap("Provider not found"),
-    });
+    do_register_provider(provider)
 }
 
 #[update(guard = "require_register_provider")]
 #[candid_method]
 fn unregister_provider(provider_id: u64) {
-    PROVIDERS.with(|p| {
-        if let Some(provider) = p.borrow().get(&provider_id) {
-            if provider.owner == ic_cdk::caller() || is_authorized(Auth::Admin) {
-                p.borrow_mut().remove(&provider_id);
-            } else {
-                ic_cdk::trap("Not authorized");
-            }
-        }
-    });
+    do_unregister_provider(provider_id)
+}
+
+#[update(guard = "require_register_provider")]
+#[candid_method]
+fn update_provider(provider: UpdateProvider) {
+    do_update_provider(provider)
 }
 
 #[query(guard = "require_register_provider")]
@@ -338,6 +279,10 @@ fn initialize() {
         metadata.open_rpc_access = DEFAULT_OPEN_RPC_ACCESS;
         m.borrow_mut().set(metadata).unwrap();
     });
+
+    for provider in get_default_providers() {
+        do_register_provider(provider);
+    }
 }
 
 #[cfg(not(any(target_arch = "wasm32", test)))]
